LayoutTests/imported/w3c/web-platform-tests/html/dom/elements/global-attributes/cdata-dir_auto-expected.txt

اختبر


PASS Content of CDATA is ignored for dir=auto in html document
PASS Text in CDATASection is considered when determining auto directionality
PASS Directionality is updated when removing CDATASection
PASS Directionality is updated when changing text of CDATASection
